\name{neg_feature_matrix10301}
\alias{neg_feature_matrix10301}
\docType{data}
\title{
control data for plate 10301
}
\description{
This data set contains all filtered feature data for the control objects of plate 10301.
The data was feature wise scaled from 0 to 1.
}
\usage{data(neg_feature_matrix10301)}
\format{
  A data frame with x observations on the following 8 variables.
  \describe{
    \item{\code{Nuclear Area}}{a numeric vector}
    \item{\code{Nuclear Extent}}{a numeric vector}
    \item{\code{max. Oct4 Intensity}}{a numeric vector}
    \item{\code{max. Hoechst Intensity}}{a numeric vector}
    \item{\code{ø Hoechst Intensity}}{a numeric vector}
    \item{\code{ø Oct4 Intensity}}{a numeric vector}
    \item{\code{Distance from first neighbor}}{a numeric vector}
    \item{\code{Number of Neighbors}}{a numeric vector}
  }
}
\examples{
data(neg_feature_matrix10301)
print(negmatrix)
}
\keyword{datasets}
