\name{origin10301}
\alias{origin10301}
\docType{data}
\title{
origin of plate 10301
}
\description{
This dataset contains the well numbers of origin of each object on plate 10301.
}
\usage{data(origin10301)}
\format{
  A vector with numberof cells observations on the following variable.
  \describe{
    \item{\code{1}}{a numeric vector}
  }
}
\examples{
data(origin10301)
print(origin)
}
\keyword{datasets}
