\name{origin10302}
\alias{origin10302}
\docType{data}
\title{
origin of plate 10302
}
\description{
This dataset contains the well numbers of origin of each object on plate 10302.
}
\usage{data(origin10302)}
\format{
  A vector with numberof cells observations on the following variable.
  \describe{
    \item{\code{1}}{a numeric vector}
  }
}
\examples{
data(origin10302)
print(origin)
}
\keyword{datasets}
