\name{origin10303}
\alias{origin10303}
\docType{data}
\title{
origin of plate 10303
}
\description{
This dataset contains the well numbers of origin of each object on plate 10303.
}
\usage{data(origin10303)}
\format{
  A vector with numberof cells observations on the following variable.
  \describe{
    \item{\code{1}}{a numeric vector}
  }
}
\examples{
data(origin10303)
print(origin)
}
\keyword{datasets}
