\name{origin10401}
\alias{origin10401}
\docType{data}
\title{
origin of plate 10401
}
\description{
This dataset contains the well numbers of origin of each object on plate 10401.
}
\usage{data(origin10401)}
\format{
  A vector with numberof cells observations on the following variable.
  \describe{
    \item{\code{1}}{a numeric vector}
  }
}
\examples{
data(origin10401)
print(origin)
}
\keyword{datasets}
