\name{origin10402}
\alias{origin10402}
\docType{data}
\title{
origin of plate 10402
}
\description{
This dataset contains the well numbers of origin of each object on plate 10402.
}
\usage{data(origin10402)}
\format{
  A vector with numberof cells observations on the following variable.
  \describe{
    \item{\code{1}}{a numeric vector}
  }
}
\examples{
data(origin10402)
print(origin)
}
\keyword{datasets}
