\name{origin10403}
\alias{origin10403}
\docType{data}
\title{
origin of plate 10403
}
\description{
This dataset contains the well numbers of origin of each object on plate 10403.
}
\usage{data(origin10403)}
\format{
  A vector with numberof cells observations on the following variable.
  \describe{
    \item{\code{1}}{a numeric vector}
  }
}
\examples{
data(origin10403)
print(origin)
}
\keyword{datasets}
