\name{raw_data_sample_10302}
\alias{raw_data_sample_10302}
\docType{data}
\title{
sample data for plate 10302
}
\description{
This data set contains all filtered feature data for each object of plate 10302.
The data was feature wise scaled from 0 to 1.
}
\usage{data(raw_data_sample_10302)}
\format{
  A data frame with x observations on the following 8 variables.
  \describe{
    \item{\code{AreaShape_Area}}{a numeric vector}
    \item{\code{AreaShape_Extent}}{a numeric vector}
    \item{\code{Intensity_MaxIntensity_OrigGreen}}{a numeric vector}
    \item{\code{Intensity_MaxIntensity_OrigBlue}}{a numeric vector}
    \item{\code{Intensity_MeanIntensity_OrigBlue}}{a numeric vector}
    \item{\code{Intensity_MeanIntensity_OrigGreen}}{a numeric vector}
    \item{\code{Neighbors_FirstClosestDistance_Expanded}}{a numeric vector}
    \item{\code{Neighbors_NumberOfNeighbors_Expanded}}{a numeric vector}
  }
}
\examples{
data(raw_data_sample_10302)
}
\keyword{datasets}
