\name{wellnames10401}
\alias{wellnames10401}
\docType{data}
\title{
Wellnames of plate 10401
}
\description{
This dataset contains the well count, well pace and RNAi reagent for each treatment on plate 10401.
}
\usage{data(wellnames10401)}
\format{
  A data frame with 352 observations on the following 3 variables.
  \describe{
    \item{\code{1}}{a numeric vector}
    \item{\code{2}}{a character vector}
    \item{\code{3}}{a character vector}
  }
}
\details{
the firt column contains the well numbers see(origin)
the second column contains the well names see(origin)
the third column contains the RNAi eagent used in that well
}
\examples{
data(wellnames10401)
print(b)
}
\keyword{datasets}
